Feature: Add Task class to represent tasks
